+++ qdax/core/containers/mome_repertoire.py
@@ -9,12 +9,10 @@ from typing import Any, Tuple
 
 import jax
 import jax.numpy as jnp
+
 from qdax.core.containers.repertoire import MapElitesRepertoire, get_cells_indices
 from qdax.types import Centroid, Descriptor, Fitness, Genotype, RNGKey
-from qdax.utils.mome_utils import (
-    sample_in_masked_pareto_front,
-    update_masked_pareto_front,
-)
+from qdax.utils.mome_utils import compute_masked_pareto_front
 
 
 class MOMERepertoire(MapElitesRepertoire):
@@ -47,6 +45,29 @@ class MOMERepertoire(MapElitesRepertoire):
         """
         first_leaf = jax.tree_leaves(self.genotypes)[0]
         return int(first_leaf.shape[0] * first_leaf.shape[1])
+
+    @partial(jax.jit, static_argnames=("num_samples",))
+    def _sample_in_masked_pareto_front(
+        self,
+        pareto_front_genotypes: Genotype,
+        mask: jnp.ndarray,
+        random_key: RNGKey,
+    ) -> Genotype:
+        """
+        Sample num_samples elements in masked pareto front.
+
+        Note: do not retrieve a random key because this function
+        is to be vmapped. The public method that uses this function
+        will return a random key
+        """
+        p = (1.0 - mask) / jnp.sum(1.0 - mask)
+
+        genotype_sample = jax.tree_map(
+            lambda x: jax.random.choice(random_key, x, shape=(1,), p=p),
+            pareto_front_genotypes,
+        )
+
+        return genotype_sample
 
     @partial(jax.jit, static_argnames=("num_samples",))
     def sample(
@@ -82,9 +103,7 @@ class MOMERepertoire(MapElitesRepertoire):
         pareto_front_genotypes = jax.tree_map(lambda x: x[cells_idx], self.genotypes)
 
         # prepare second sampling function
-        sample_in_fronts = jax.vmap(
-            partial(sample_in_masked_pareto_front, num_samples=1)
-        )
+        sample_in_fronts = jax.vmap(self._sample_in_masked_pareto_front)
 
         # sample genotypes from the pareto front
         random_key, subkey = jax.random.split(random_key)
@@ -99,6 +118,101 @@ class MOMERepertoire(MapElitesRepertoire):
         sampled_genotypes = jax.tree_map(lambda x: x.squeeze(), sampled_genotypes)
 
         return sampled_genotypes, random_key
+
+    @jax.jit
+    def _update_masked_pareto_front(
+        self,
+        pareto_front_fitness: Fitness,
+        pareto_front_genotypes: Genotype,
+        pareto_front_descriptors: Descriptor,
+        mask: jnp.ndarray,
+        new_batch_of_criteria: Fitness,
+        new_batch_of_genotypes: Genotype,
+        new_batch_of_descriptors: Descriptor,
+        new_mask: jnp.ndarray,
+    ) -> Tuple[Fitness, Genotype, Descriptor, jnp.ndarray]:
+        """
+        Takes a fixed size pareto front, its mask and new points to add.
+        Returns updated front and mask.
+        """
+        # get dimensions
+        batch_size = new_batch_of_criteria.shape[0]
+        num_criteria = new_batch_of_criteria.shape[1]
+
+        pareto_front_len = pareto_front_fitness.shape[0]
+
+        first_leaf = jax.tree_leaves(new_batch_of_genotypes)[0]
+        genotypes_dim = first_leaf.shape[1]
+
+        descriptors_dim = new_batch_of_descriptors.shape[1]
+
+        # gather all data
+        cat_mask = jnp.concatenate([mask, new_mask], axis=-1)
+        cat_f = jnp.concatenate([pareto_front_fitness, new_batch_of_criteria], axis=0)
+        cat_genotypes = jax.tree_map(
+            lambda x, y: jnp.concatenate([x, y], axis=0),
+            pareto_front_genotypes,
+            new_batch_of_genotypes,
+        )
+        cat_descriptors = jnp.concatenate(
+            [pareto_front_descriptors, new_batch_of_descriptors], axis=0
+        )
+
+        # get new front
+        cat_bool_front = compute_masked_pareto_front(
+            batch_of_criteria=cat_f, mask=cat_mask
+        )
+
+        # get corresponding indices
+        indices = (
+            jnp.arange(start=0, stop=pareto_front_len + batch_size) * cat_bool_front
+        )
+        indices = indices + ~cat_bool_front * (batch_size + pareto_front_len - 1)
+        indices = jnp.sort(indices)
+
+        # get new fitness, genotypes and descriptors
+        new_front_fitness = jnp.take(cat_f, indices, axis=0)
+        new_front_genotypes = jax.tree_map(
+            lambda x: jnp.take(x, indices, axis=0), cat_genotypes
+        )
+        new_front_descriptors = jnp.take(cat_descriptors, indices, axis=0)
+
+        # compute new mask
+        num_front_elements = jnp.sum(cat_bool_front)
+        new_mask_indices = jnp.arange(start=0, stop=batch_size + pareto_front_len)
+        new_mask_indices = (num_front_elements - new_mask_indices) > 0
+
+        new_mask = jnp.where(
+            new_mask_indices,
+            jnp.ones(shape=batch_size + pareto_front_len, dtype=bool),
+            jnp.zeros(shape=batch_size + pareto_front_len, dtype=bool),
+        )
+
+        fitness_mask = jnp.repeat(
+            jnp.expand_dims(new_mask, axis=-1), num_criteria, axis=-1
+        )
+        new_front_fitness = new_front_fitness * fitness_mask
+        new_front_fitness = new_front_fitness[: len(pareto_front_fitness), :]
+
+        genotypes_mask = jnp.repeat(
+            jnp.expand_dims(new_mask, axis=-1), genotypes_dim, axis=-1
+        )
+        new_front_genotypes = jax.tree_map(
+            lambda x: x * genotypes_mask, new_front_genotypes
+        )
+        new_front_genotypes = jax.tree_map(
+            lambda x: x[: len(pareto_front_fitness), :], new_front_genotypes
+        )
+
+        descriptors_mask = jnp.repeat(
+            jnp.expand_dims(new_mask, axis=-1), descriptors_dim, axis=-1
+        )
+        new_front_descriptors = new_front_descriptors * descriptors_mask
+        new_front_descriptors = new_front_descriptors[: len(pareto_front_fitness), :]
+
+        new_mask = ~new_mask[: len(pareto_front_fitness)]
+
+        return new_front_fitness, new_front_genotypes, new_front_descriptors, new_mask
 
     @jax.jit
     def add(
@@ -151,7 +265,7 @@ class MOMERepertoire(MapElitesRepertoire):
                 cell_genotype,
                 cell_descriptor,
                 cell_mask,
-            ) = update_masked_pareto_front(
+            ) = self._update_masked_pareto_front(
                 pareto_front_fitness=cell_fitness.squeeze(),
                 pareto_front_genotypes=cell_genotype.squeeze(),
                 pareto_front_descriptors=cell_descriptor.squeeze(),
@@ -214,7 +328,8 @@ class MOMERepertoire(MapElitesRepertoire):
         Args:
             genotypes: initial genotypes, pytree in which leaves
                 have shape (batch_size, num_features)
-            fitnesses: fitness of the initial genotypes of shape (batch_size, num_criteria)
+            fitnesses: fitness of the initial genotypes of shape:
+                (batch_size, num_criteria)
             descriptors: descriptors of the initial genotypes
                 of shape (batch_size, num_descriptors)
             centroids: tesselation centroids of shape (batch_size, num_descriptors)
@@ -259,3 +374,17 @@ class MOMERepertoire(MapElitesRepertoire):
         new_repertoire = repertoire.add(genotypes, descriptors, fitnesses)
 
         return new_repertoire  # type: ignore
+
+
+def compute_global_pareto_front(
+    grid: MOMERepertoire,
+) -> Tuple[jnp.ndarray, jnp.ndarray]:
+    """
+    Merge all the pareto fronts of the MOME grid into a single one called global.
+    """
+    fitnesses = jnp.concatenate(grid.fitnesses, axis=0)
+    mask = jnp.any(fitnesses == -jnp.inf, axis=-1)
+    pareto_bool = compute_masked_pareto_front(fitnesses, mask)
+    pareto_front = fitnesses - jnp.inf * (~jnp.array([pareto_bool, pareto_bool]).T)
+
+    return pareto_front, pareto_bool
